back_django/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
black==22.1.0
click==8.0.3
Django==4.0.2
django-debug-toolbar==3.2.4
django-environ==0.8.1
django-phonenumber-field==6.1.0
flake8==4.0.1
install==1.3.5
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
phonenumberslite==8.12.43
platformdirs==2.5.0
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
python-dotenv==0.19.2
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.1.1
